+++ db/migrate/20240710040009_create_trips.rb
@@ -1,27 +1,26 @@
 # db/migrate/xxxxxxxxxx_create_trips.rb
 class CreateTrips < ActiveRecord::Migration[7.1]
   def change
-    # Drop the table if it already exists
-    if ActiveRecord::Base.connection.table_exists?(:trips)
-      drop_table :trips
+    reversible do |dir|
+      dir.up do
+        drop_table :trips if table_exists?(:trips)
+
+        create_table :trips, id: :string, primary_key: :id do |t|
+          t.string :user_driver_id, null: false
+          t.string :ride_id, null: false
+          t.decimal :commute_minutes, precision: 10, scale: 2
+          t.decimal :commute_miles, precision: 10, scale: 2
+          t.decimal :total_minutes, precision: 10, scale: 2
+          t.decimal :total_hours, precision: 10, scale: 2
+          t.decimal :total_miles, precision: 10, scale: 2
+          t.decimal :earnings, precision: 10, scale: 2
+          t.timestamps
+        end
+
+        add_index :trips, :id, unique: true
+        add_foreign_key :trips, :user_drivers, column: :user_driver_id, primary_key: :id
+        add_foreign_key :trips, :rides, column: :ride_id, primary_key: :id
+      end
     end
-
-    create_table :trips, id: false do |t|
-      t.string :id, null: false, primary_key: true
-      t.string :driver_id, null: false
-      t.string :ride_id, null: false
-      t.decimal :commute_minutes, precision: 10, scale: 2
-      t.decimal :commute_miles, precision: 10, scale: 2
-      t.decimal :total_minutes, precision: 10, scale: 2
-      t.decimal :total_hours, precision: 10, scale: 2
-      t.decimal :total_miles, precision: 10, scale: 2
-      t.decimal :earnings, precision: 10, scale: 2
-
-      t.timestamps
-    end
-
-    add_index :trips, :id, unique: true
-    add_foreign_key :trips, :drivers, column: :driver_id, primary_key: :id
-    add_foreign_key :trips, :rides, column: :ride_id, primary_key: :id
   end
 end
